LayoutTests/imported/w3c/web-platform-tests/workers/interfaces/WorkerGlobalScope/onerror/exception-in-onerror-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: y

PASS onerror, "not handled" with an error in the onerror function
